Split common js into core files
