+++ admin/js/core/index.js
@@ -1,0 +1,7 @@
+/**
+ * Shared core functionalities and setup
+ */
+import './qtranslatex';
+
+export * from './dom';
+export * from './qblocks';
